Major update: no functionality, but massive code restructuring
So my code structure hasn't been the best. Hopefully this new separation
will make things more readable:

-There was some disgusting inter-class coupling; now all model-view communication
is handled by the controller package.
-Some gui elements used to be their own listeners. I'm now trying to make
the gui as dumb as possible.
-When the model is updated, the observing view will make changes accordingly.
The model does not know about the controller or view.

My next priority is an efficiency boost. I'll get working on that tomorrow.